requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
PyJWT==2.11.0
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3
